+++ src/fields/Lightswitch.php
@@ -172,16 +172,15 @@ class Lightswitch extends Field implements PreviewableFieldInterface, SortableFi
     /**
      * @inheritdoc
      */
-    public function modifyElementsQuery(ElementQueryInterface $query, $value)
+    public function modifyElementsQuery(ElementQueryInterface $query, $value): void
     {
         /** @var ElementQuery $query */
         if ($value === null) {
-            return null;
+            return;
         }
 
         $column = ElementHelper::fieldColumnFromField($this);
         $query->subQuery->andWhere(Db::parseBooleanParam("content.$column", $value, (bool)$this->default));
-        return null;
     }
 
     /**